# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  picture?: string | null;
  /** Whether background auto-labeling is enabled for this user. */
  autoLabelEnabled?: boolean;
  /** Whether daily unread label digests are emailed to the user. */
  dailyDigestEnabled?: boolean;
}

export interface UpdateSettingsInput {
  autoLabelEnabled?: boolean;
  dailyDigestEnabled?: boolean;
}

export interface ConnectedAccount {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  picture: string | null;
  isPrimary: boolean;
  isActive: boolean;
}

export interface SwitchAccountInput {
  accountId: number;
}

export interface Label {
  id: string;
  name: string;
  /**
     * A hex color string, e.g. "#6366f1"
     * @nullable
     */
  color: string | null;
  /** @nullable */
  description?: string | null;
  isSystem: boolean;
  emailCount: number;
}

export interface LabelInput {
  /** @minLength 1 */
  name: string;
  color: string;
  description?: string;
}

export interface LabelUpdate {
  /** @minLength 1 */
  name?: string;
  color?: string;
  /** @nullable */
  description?: string | null;
}

export interface Email {
  id: string;
  sender: string;
  senderEmail: string;
  subject: string;
  snippet: string;
  /** Plain-text rendering of the email body (HTML stripped). Used for previews and AI context. */
  body: string;
  /** Sanitized HTML body for rich display. Empty string when the email has no HTML part. Remote image sources are held in data-blocked-src until the user displays images. */
  bodyHtml: string;
  receivedAt: string;
  isRead: boolean;
  isStarred: boolean;
  labels: Label[];
  /** True when the sanitized HTML has remote images that were held back (blocked by default). */
  hasRemoteImages: boolean;
  /**
     * One-click/managed https unsubscribe target from List-Unsubscribe, if any.
     * @nullable
     */
  unsubscribeUrl: string | null;
  /**
     * A mailto unsubscribe target from List-Unsubscribe, if any.
     * @nullable
     */
  unsubscribeMailto: string | null;
}

export interface EmailPage {
  emails: Email[];
  /**
     * Cursor to fetch the next page; null when the window is exhausted.
     * @nullable
     */
  nextPageToken: string | null;
}

export interface EmailUpdate {
  isRead?: boolean;
  isStarred?: boolean;
}

export interface EmailLabelInput {
  labelIds: string[];
}

export type BulkLabelInputAction = typeof BulkLabelInputAction[keyof typeof BulkLabelInputAction];


export const BulkLabelInputAction = {
  add: 'add',
  remove: 'remove',
} as const;

export interface BulkLabelInput {
  emailIds: string[];
  labelId: string;
  action: BulkLabelInputAction;
}

export type BulkActionInputAction = typeof BulkActionInputAction[keyof typeof BulkActionInputAction];


export const BulkActionInputAction = {
  archive: 'archive',
  trash: 'trash',
  untrash: 'untrash',
  spam: 'spam',
  markRead: 'markRead',
  markUnread: 'markUnread',
  star: 'star',
  unstar: 'unstar',
} as const;

export interface BulkActionInput {
  emailIds: string[];
  action: BulkActionInputAction;
}

export interface SendEmailInput {
  /** Recipient address(es), comma-separated. */
  to: string;
  cc?: string;
  subject: string;
  /** Plain-text message body. */
  body: string;
  /** Source message id when replying/forwarding; threads the message. */
  inReplyToId?: string;
}

export interface SendEmailResult {
  id: string;
  /** @nullable */
  threadId: string | null;
}

/**
 * posted = one-click POST done; open = client should open url; none = no target.
 */
export type UnsubscribeResultStatus = typeof UnsubscribeResultStatus[keyof typeof UnsubscribeResultStatus];


export const UnsubscribeResultStatus = {
  posted: 'posted',
  open: 'open',
  none: 'none',
} as const;

export interface UnsubscribeResult {
  /** posted = one-click POST done; open = client should open url; none = no target. */
  status: UnsubscribeResultStatus;
  /** @nullable */
  url: string | null;
}

export interface LabelSuggestion {
  /**
     * Set when the suggestion maps to an existing label
     * @nullable
     */
  labelId?: string | null;
  name: string;
  /** @nullable */
  color?: string | null;
  /** 0 to 1 confidence score */
  confidence: number;
  reason: string;
  /** True when this is a brand-new label idea, not an existing label */
  isNew: boolean;
}

export interface AutoLabelInput {
  /** Specific emails to label. If omitted, all unlabeled emails are processed. */
  emailIds?: string[];
}

export interface AutoLabelItem {
  emailId: string;
  appliedLabelIds: string[];
}

export interface AutoLabelResult {
  processed: number;
  labeled: number;
  items: AutoLabelItem[];
}

export interface GroupSuggestInput {
  /** Emails to cluster. If omitted, all unlabeled emails are used. */
  emailIds?: string[];
}

export interface EmailGroup {
  suggestedLabel: string;
  suggestedColor: string;
  reason: string;
  emailIds: string[];
}

export type DigestInputView = typeof DigestInputView[keyof typeof DigestInputView];


export const DigestInputView = {
  all: 'all',
  unlabeled: 'unlabeled',
  starred: 'starred',
  unread: 'unread',
} as const;

export interface DigestInput {
  /** Scope the digest to emails with this label. */
  labelId?: string;
  view?: DigestInputView;
  /** Restrict to unread emails. */
  onlyUnread?: boolean;
}

export interface DigestItem {
  id: string;
  subject: string;
  sender: string;
  summary: string;
}

export interface DigestResult {
  summary: string;
  count: number;
  items: DigestItem[];
}

export interface LabelStat {
  id: string;
  name: string;
  /** @nullable */
  color: string | null;
  count: number;
}

export interface Stats {
  totalEmails: number;
  labeledCount: number;
  unlabeledCount: number;
  unreadCount: number;
  starredCount: number;
  labelCount: number;
  labelBreakdown: LabelStat[];
}

export type ListEmailsParams = {
/**
 * Filter to emails that have this Gmail label id
 */
labelId?: string;
/**
 * Predefined filter view
 */
view?: ListEmailsView;
/**
 * Gmail search query across subject, sender and body
 */
search?: string;
/**
 * Opaque cursor from a previous response's nextPageToken to fetch the next page
 */
pageToken?: string;
};

export type ListEmailsView = typeof ListEmailsView[keyof typeof ListEmailsView];


export const ListEmailsView = {
  all: 'all',
  unlabeled: 'unlabeled',
  starred: 'starred',
  unread: 'unread',
} as const;