# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /emails:
    get:
      operationId: listEmails
      tags: [emails]
      summary: List emails
      description: List the authenticated user's Gmail messages with optional filtering by label, view, and search. Returns one bounded page plus a cursor for the next.
      parameters:
        - name: labelId
          in: query
          required: false
          schema:
            type: string
          description: Filter to emails that have this Gmail label id
        - name: view
          in: query
          required: false
          schema:
            type: string
            enum: [all, unlabeled, starred, unread]
          description: Predefined filter view
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Gmail search query across subject, sender and body
        - name: pageToken
          in: query
          required: false
          schema:
            type: string
          description: Opaque cursor from a previous response's nextPageToken to fetch the next page
      responses:
        "200":
          description: A page of emails
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailPage"
  /emails/{id}:
    get:
      operationId: getEmail
      tags: [emails]
      summary: Get a single email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateEmail
      tags: [emails]
      summary: Update an email's read/starred state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailUpdate"
      responses:
        "200":
          description: Updated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
  /emails/{id}/labels:
    put:
      operationId: setEmailLabels
      tags: [emails]
      summary: Replace the set of user labels on an email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailLabelInput"
      responses:
        "200":
          description: Updated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
...
  /ai/suggest-labels/{id}:
    post:
      operationId: suggestEmailLabels
      tags: [ai]
      summary: AI-suggested labels for a single email
      description: Uses AI to analyze the email and suggest relevant labels from existing labels, plus brand-new label ideas.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Suggestions for the email
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LabelSuggestion"
  /ai/auto-label:
    post:
      operationId: autoLabelEmails
      tags: [ai]
      summary: Automatically label emails with AI
      description: Runs AI labeling over the given emails (or the whole inbox) and applies the best matching existing labels.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoLabelInput"
      responses:
        "200":
          description: The result of the auto-labeling run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoLabelResult"
  /ai/group-suggestions:
    post:
      operationId: suggestEmailGroups
      tags: [ai]
      summary: AI-suggested groups of similar emails
      description: Uses AI to cluster similar unlabeled emails and propose a common label for each cluster.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupSuggestInput"
      responses:
        "200":
          description: Suggested groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailGroup"
  /ai/digest:
    post:
      operationId: digestEmails
      tags: [ai]
      summary: AI digest of a set of emails
      description: Summarizes a scoped set of emails (by view/label, optionally unread-only) into an overview plus per-email one-liners.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DigestInput"
      responses:
        "200":
          description: The digest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigestResult"
...
  /accounts/switch:
    post:
      operationId: switchAccount
      tags: [accounts]
      summary: Switch the active connected account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchAccountInput"
      responses:
        "200":
          description: The updated account list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConnectedAccount"
  /accounts/{id}:
    delete:
      operationId: unlinkAccount
      tags: [accounts]
      summary: Unlink a connected account
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The remaining account list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ConnectedAccount"
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ConnectedAccount:
      type: object
      required: [id, email, name, picture, isPrimary, isActive]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        picture:
          type: ["string", "null"]
        isPrimary:
          type: boolean
        isActive:
          type: boolean
    SwitchAccountInput:
      type: object
      required: [accountId]
      properties:
        accountId:
          type: integer
    Label:
      type: object
      required: [id, name, color, isSystem, emailCount]
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: ["string", "null"]
          description: A hex color string, e.g. "#6366f1"
        description:
          type: ["string", "null"]
        isSystem:
          type: boolean
        emailCount:
          type: integer
...
    Email:
      type: object
      required:
        - id
        - sender
        - senderEmail
        - subject
        - snippet
        - body
        - bodyHtml
        - receivedAt
        - isRead
        - isStarred
        - labels
        - hasRemoteImages
        - unsubscribeUrl
        - unsubscribeMailto
      properties:
        id:
          type: string
        sender:
          type: string
        senderEmail:
          type: string
        subject:
          type: string
        snippet:
          type: string
        body:
          type: string
          description: Plain-text rendering of the email body (HTML stripped). Used for previews and AI context.
        bodyHtml:
          type: string
          description: Sanitized HTML body for rich display. Empty string when the email has no HTML part. Remote image sources are held in data-blocked-src until the user displays images.
        receivedAt:
          type: string
          format: date-time
        isRead:
          type: boolean
        isStarred:
          type: boolean
        labels:
          type: array
          items:
            $ref: "#/components/schemas/Label"
        hasRemoteImages:
          type: boolean
          description: True when the sanitized HTML has remote images that were held back (blocked by default).
        unsubscribeUrl:
          type: ["string", "null"]
          description: One-click/managed https unsubscribe target from List-Unsubscribe, if any.
        unsubscribeMailto:
          type: ["string", "null"]
          description: A mailto unsubscribe target from List-Unsubscribe, if any.
    EmailPage:
      type: object
      required: [emails, nextPageToken]
      properties:
        emails:
          type: array
          items:
            $ref: "#/components/schemas/Email"
        nextPageToken:
          type: ["string", "null"]
          description: Cursor to fetch the next page; null when the window is exhausted.
    EmailUpdate:
      type: object
      properties:
        isRead:
          type: boolean
        isStarred:
          type: boolean
    EmailLabelInput:
      type: object
      required: [labelIds]
      properties:
        labelIds:
          type: array
          items:
            type: string
...
    LabelSuggestion:
      type: object
      required: [name, confidence, reason, isNew]
      properties:
        labelId:
          type: ["string", "null"]
          description: Set when the suggestion maps to an existing label
        name:
          type: string
        color:
          type: ["string", "null"]
        confidence:
          type: number
          description: 0 to 1 confidence score
        reason:
          type: string
        isNew:
          type: boolean
          description: True when this is a brand-new label idea, not an existing label
    AutoLabelInput:
      type: object
      properties:
        emailIds:
          type: array
          items:
            type: string
          description: Specific emails to label. If omitted, all unlabeled emails are processed.
    AutoLabelItem:
      type: object
      required: [emailId, appliedLabelIds]
      properties:
        emailId:
          type: string
        appliedLabelIds:
          type: array
          items:
            type: string
    AutoLabelResult:
      type: object
      required: [processed, labeled, items]
      properties:
        processed:
          type: integer
        labeled:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/AutoLabelItem"
    GroupSuggestInput:
      type: object
      properties:
        emailIds:
          type: array
          items:
            type: string
          description: Emails to cluster. If omitted, all unlabeled emails are used.
    EmailGroup:
      type: object
      required: [suggestedLabel, suggestedColor, reason, emailIds]
      properties:
        suggestedLabel:
          type: string
        suggestedColor:
          type: string
        reason:
          type: string
        emailIds:
          type: array
          items:
            type: string
    DigestInput:
      type: object
      properties:
        labelId:
          type: string
          description: Scope the digest to emails with this label.
        view:
          type: string
          enum: [all, unlabeled, starred, unread]
        onlyUnread:
          type: boolean
          description: Restrict to unread emails.
    DigestItem:
      type: object
      required: [id, subject, sender, summary]
      properties:
        id:
          type: string
        subject:
          type: string
        sender:
          type: string
        summary:
          type: string
    DigestResult:
      type: object
      required: [summary, count, items]
      properties:
        summary:
          type: string
        count:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/DigestItem"
...